app/requirements.txt
cassandra-driver
pyspark==3.5.4
pathvalidate
